Decompose widget_creators.go and TypeScript widget files
Split large files into smaller modules to improve LLM token efficiency
and ts-node transpilation speed:

Go (bridge/):
- widget_creators.go (4301 lines) -> 5 files by category:
  - widget_creators_inputs.go (Button, Entry, Checkbox, etc.)
  - widget_creators_display.go (Label, Image, Icon, etc.)
  - widget_creators_containers.go (VBox, HBox, Tabs, etc.)
  - widget_creators_canvas.go (Canvas primitives)
  - widget_creators_complex.go (Tree, Table, List, etc.)

TypeScript (src/widgets/):
- containers.ts -> 4 sub-modules (box, layout, tabs, advanced)
- display.ts -> 4 sub-modules (basic, progress, data, text)
- inputs.ts -> 3 sub-modules (text, selection, range)

All original modules now re-export from sub-modules for backward
compatibility.
